app/requirements.txt
# ==============================================================================
# PYTHON DEPENDENCIES FOR WARMTECHECK APPLICATION
# ==============================================================================

# Azure Functions Runtime
azure-functions>=1.11.0

# Azure SDK Dependencies
azure-identity>=1.12.0
azure-storage-blob>=12.14.0

# HTTP Client
requests>=2.28.0

# Development Dependencies (optional)
# Uncomment for development
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.991

# ==============================================================================
# PACKAGE VERSIONS EXPLANATION
# ==============================================================================
# azure-functions: Azure Functions runtime for Python
# azure-identity: Azure authentication (DefaultAzureCredential)
# azure-storage-blob: Azure Blob Storage client
# requests: HTTP library for API calls (OpenMeteo)
# ==============================================================================
